+++ Expiry/Prepare.cs
@@ -10,14 +10,11 @@ namespace Expiry
         public static async Task Stage(string connectionString, string destination)
         {
             var client = new ManagementClient(connectionString);
-            if (await client.QueueExistsAsync(destination))
-            {
-                await client.DeleteQueueAsync(destination);
-            }
+            if (await client.QueueExistsAsync(destination)) await client.DeleteQueueAsync(destination);
 
             var description = new QueueDescription(destination)
             {
-                MaxDeliveryCount = int.MaxValue,
+                MaxDeliveryCount = int.MaxValue
             };
             await client.CreateQueueAsync(description);
 
@@ -27,15 +24,15 @@ namespace Expiry
         public static async Task SimulateActiveReceiver(QueueClient client)
         {
             client.RegisterMessageHandler(
-            async (message, token) =>
-            {
-                await client.AbandonAsync(message.SystemProperties.LockToken);
-                await Task.Delay(2000, token);
-            },
-            new MessageHandlerOptions(exception => Task.CompletedTask)
-            {
-                AutoComplete = false,
-            });
+                async (message, token) =>
+                {
+                    await client.AbandonAsync(message.SystemProperties.LockToken);
+                    await Task.Delay(2000, token);
+                },
+                new MessageHandlerOptions(exception => Task.CompletedTask)
+                {
+                    AutoComplete = false
+                });
 
             await Task.Delay(TimeSpan.FromSeconds(15));
         }
