+++ Session/Prepare.cs
@@ -8,15 +8,12 @@ namespace Session
         public static async Task Stage(string connectionString, string destination)
         {
             var client = new ManagementClient(connectionString);
-            if (await client.QueueExistsAsync(destination))
-            {
-                await client.DeleteQueueAsync(destination);
-            }
+            if (await client.QueueExistsAsync(destination)) await client.DeleteQueueAsync(destination);
             var queueDescription = new QueueDescription(destination)
-                {
-                    RequiresSession = true
-                };
-                await client.CreateQueueAsync(queueDescription);
+            {
+                RequiresSession = true
+            };
+            await client.CreateQueueAsync(queueDescription);
             await client.CloseAsync();
         }
     }
